backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
sqlalchemy==2.0.35
alembic==1.13.2
httpx==0.27.2
prometheus_client==0.21.0
pyyaml==6.0.2
scikit-learn==1.5.1
numpy==2.1.0
regex==2024.7.24
